Fix float64 type mismatch for Strava time fields
Strava API returns moving_time and elapsed_time as floats in stats
responses (e.g. 11446.0). Changed all time fields to float64 across
Activity, ActivityTotal, and format.Duration to handle this correctly.

+++ internal/api/activities.go
@@ -12,8 +12,8 @@ type Activity struct {
 	Type               string  `json:"type"`
 	SportType          string  `json:"sport_type"`
 	Distance           float64 `json:"distance"`
-	MovingTime         int     `json:"moving_time"`
-	ElapsedTime        int     `json:"elapsed_time"`
+	MovingTime         float64 `json:"moving_time"`
+	ElapsedTime        float64 `json:"elapsed_time"`
 	TotalElevationGain float64 `json:"total_elevation_gain"`
 	StartDate          string  `json:"start_date"`
 	StartDateLocal     string  `json:"start_date_local"`
